requirements.txt
fastapi==0.115.5
openai==1.56.1
python-dotenv==1.0.1
uvicorn==0.32.1
python-multipart==0.0.19